Refactor classes to use inheritance
