+++ 4.5_dry_up_with_inheritance/src/main/java/com/thinkful/drills/Knife.java
@@ -1,11 +1,7 @@
 package com.thinkful.drills;
 
-public class Knife {
-  private int strength;
-  private int damage;
-  private double cost;
-  private double weight;
-  private String name;
+public class Knife extends Weapon {
+
   private double sharpness;
 
   /**
@@ -60,53 +56,10 @@ public class Knife {
         double weight, 
         double cost, 
         double sharpness) {
-    this.setName(name);
-    this.setDamage(damage);
-    this.setStrength(strength);
-    this.setWeight(weight);
-    this.setCost(cost);
+    super(name, damage, strength, weight, cost);
     this.setSharpness(sharpness);
   }
 
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getName() {
-    return this.name;
-  }
-
-  public void setDamage(int damage) {
-    this.damage = damage;
-  }
-
-  public int getDamage() {
-    return this.damage;
-  }
-
-  public void setStrength(int strength) {
-    this.strength = strength;
-  }
-
-  public int getStrength() {
-    return this.strength;
-  }
-
-  public void setWeight(double weight) {
-    this.weight = weight;
-  }
-
-  public double getWeight() {
-    return this.weight;
-  }
-
-  public void setCost(double cost) {
-    this.cost = cost;
-  }
-
-  public double getCost() {
-    return this.cost;
-  }
 
   public void setSharpness(double sharpness) {
     this.sharpness = sharpness;
@@ -114,19 +67,5 @@ public class Knife {
 
   public double getSharpness() {
     return this.sharpness;
-  }
-
-  /**
-   * Performs a knife attack on a Wizard.
-   */
-  public void attack(Wizard wizard) {
-    wizard.hit(this.getDamage());
-  }
-
-  /**
-   *  Perform a Knife attack on a dwarf.
-   */
-  public void attack(Dwarf dwarf) {
-    dwarf.hit(this.getDamage());
   }
 }
