+++ 4.5_dry_up_with_inheritance/src/main/java/com/thinkful/drills/Wizard.java
@@ -1,10 +1,7 @@
 package com.thinkful.drills;
 
-public class Wizard {
-  private String name;
-  private String description;
-  private int health;
-  private double strength;
+public class Wizard extends Character { 
+
   private double magicRating;
   
   public static final int INITIAL_HEALTH = 100;
@@ -50,43 +47,8 @@ public class Wizard {
    * Create a Wizard with name, description, health, strength and default values.
    */
   public Wizard(String name, String description, int health, double strength, double magicRating) {
-    this.setName(name);
-    this.setHealth(health);
-    this.setStrength(strength);
-    this.setDescription(description);
+    super(name, description, health, strength);
     this.setMagicRating(magicRating);
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getName() {
-    return this.name;
-  }
-
-  public void setDescription(String description) {
-    this.description = description;
-  }
-
-  public String getDescription() {
-    return this.description;
-  }
-
-  public void setHealth(int health) {
-    this.health = health;
-  }
-
-  public int getHealth() {
-    return this.health;
-  }
-
-  public void setStrength(double strength) {
-    this.strength = strength;
-  }
-
-  public double getStrength() {
-    return this.strength;
   }
 
   public void setMagicRating(double magicRating) {
@@ -95,17 +57,6 @@ public class Wizard {
 
   public double getMagicRating() {
     return this.magicRating;
-  }
-
-  /**
-   * Apply a hit to this character.
-   * A hit subtracts powerOfHit from the total health of
-   * the character. If health drops below 0 it is reset to 0.
-   * 
-   * @param powerOfHit int The amount of damage that is done to the character
-   */
-  public void hit(int powerOfHit) {
-    this.setHealth(Math.max(this.getHealth() - powerOfHit, 0));
   }
 
 }
